new void update() function
new update() allows us to delete the
excess rectangles that form a tail as well as block the snake
from leaving the screen. First, keyPressed() changed to set the
xmovement and ymovement to -1,0, or 1. By assigning posx
+= xmovement and posy += ymovement in the update function,
everytime the program updates the snake will keep moving in the
corresponding direction. Through a series of loops in update()
we can prevent posx and posy from changing when it hits an edge.
Redrawing the background() everytime the draw() function runs
removes the excess rectangles, buts hids the food. By changing
the spawnFood() to return a list of foodX and foodY, we are
able to redraw the food in the draw() function. These changes
make it more similar to the original game.